backend/requirements.txt
# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
fastapi-login
google-genai
# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
# Cloud Storage (S3)
boto3>=1.34.0
botocore>=1.34.0
# Authentication & Security
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
pydantic[email]

# AI & Image Processing

pillow==10.1.0
sentence-transformers==2.2.2
transformers==4.36.0
huggingface-hub==0.20.0
numpy==1.24.3

# Additional Services
geopy==2.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1
